refactor: move fixtures to conftest.py
fixtures were moved to conftest.py to allow for better organization
and reuse across multiple test files. This change helps in maintaining
cleaner test code and improves the overall structure of the test suite.
