+++ tests/__init__.py
@@ -15,15 +15,11 @@
 # ===============================================================================
 from alembic.config import Config
 from alembic import command
-import uuid
 
-import pytest
 from fastapi.testclient import TestClient
 
 from core.app import init_lexicon
 from main import app
-from db import *
-from db.engine import session_ctx
 
 
 def run_alembic_upgrade():
@@ -42,85 +38,6 @@ run_alembic_upgrade()
 init_lexicon()
 
 client = TestClient(app)
-
-
-@pytest.fixture(scope="session")
-def location():
-    with session_ctx() as session:
-        loc = Location(point="SRID=4326;POINT(0 0)")
-        session.add(loc)
-        session.commit()
-        session.refresh(loc)
-        yield loc
-
-        session.close()
-
-
-@pytest.fixture(scope="session")
-def thing(location):
-    with session_ctx() as session:
-        # loc = Location(point='SRID=4326;POINT(0 0)')
-        # session.add(loc)
-        # session.commit()
-        # session.refresh(loc)
-
-        wt = add_thing(
-            session,
-            {
-                "location_id": location.id,
-                "name": "Test Well",
-            },
-            "water well",
-        )
-
-        yield wt
-
-        session.close()
-
-
-@pytest.fixture(scope="session")
-def sample(thing):
-    with session_ctx() as session:
-        sample = Sample(
-            collection_timestamp="2025-01-01T00:00:00Z",
-            collection_method="manual",
-            thing_id=thing.id,
-            sample_type="groundwater",
-            sampler="Test Sampler",
-        )
-        session.add(sample)
-        session.commit()
-        yield sample
-
-        session.close()
-
-
-@pytest.fixture(scope="session")
-def sensor():
-    with session_ctx() as session:
-        sensor = Sensor(name=f"Test Sensor {uuid.uuid4()}")
-        session.add(sensor)
-        session.commit()
-        session.refresh(thing)
-        yield thing
-        session.delete(thing)
-        session.close()
-
-
-@pytest.fixture
-def sample_fixture(thing):
-    with session_ctx() as session:
-        sample = Sample(
-            thing_id=thing.id,
-            collection_timestamp="2025-01-01T00:00:00+00:00",
-            collection_method="manual",
-        )
-        session.add(sample)
-        session.commit()
-        session.refresh(sample)
-        yield thing, sample
-        session.delete(sample)
-        session.commit()
 
 
 # ============= EOF =============================================
